requirements.txt
PyMuPDF==1.23.8
sentence-transformers==2.5.1
faiss-cpu==1.7.4
google-generativeai==0.3.2
python-dotenv==1.0.1
tqdm==4.66.2
